pyproject.toml
[project]
name = "helpful-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
]
